Added Extensions
The thing now has WAY more functionality, and I added the extensions that a missed reading about earlier, and the script looks a bit messy in my opinion but I tried to make it as readable as possible. This is probably my final submission then, and hopefully it covers everything.